+++ contracts/handlers/yearn/HYVault.sol
@@ -4,7 +4,6 @@ import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
 import "../HandlerBase.sol";
 import "./IYVault.sol";
 
-
 contract HYVault is HandlerBase {
     using SafeERC20 for IERC20;
 
@@ -18,10 +17,22 @@ contract HYVault is HandlerBase {
         _updateToken(address(yVault));
     }
 
+    function depositETH(address vault, uint256 _amount) external payable {
+        IYVault yVault = IYVault(vault);
+        yVault.depositETH.value(_amount)();
+
+        _updateToken(address(yVault));
+    }
+
     function withdraw(address vault, uint256 _shares) external payable {
         IYVault yVault = IYVault(vault);
         yVault.withdraw(_shares);
 
         _updateToken(yVault.token());
+    }
+
+    function withdrawETH(address vault, uint256 _shares) external payable {
+        IYVault yVault = IYVault(vault);
+        yVault.withdrawETH(_shares);
     }
 }